weaviate_benchmarks/requirements.txt
weaviate-client>=4.9.0
numpy>=1.24
matplotlib>=3.7
